Separated the base and console classes and added an `__init__.py` to hide ugly imports.
This is due to too much logic related to managing the screen (and a future cleanup to all ANSII escape sequences).
